+++ bungee/src/main/java/com/github/games647/fastlogin/bungee/listener/PlayerConnectionListener.java
@@ -1,7 +1,7 @@
 package com.github.games647.fastlogin.bungee.listener;
 
-import com.github.games647.fastlogin.bungee.tasks.AsyncPremiumCheck;
 import com.github.games647.fastlogin.bungee.FastLoginBungee;
+import com.github.games647.fastlogin.bungee.tasks.AsyncPremiumCheck;
 import com.github.games647.fastlogin.bungee.tasks.ForceLoginTask;
 import com.github.games647.fastlogin.core.LoginSession;
 import com.github.games647.fastlogin.core.PlayerProfile;